+++ packages/kiarina-agi-base/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Default run context identifiers to `default`.
+
 ## [2.6.0] - 2026-07-03
 
 ### Changed